SoftwareEngs/SlidesShareYetkin/requirements.txt
pysimplegui==4.60.5
requests
pillow
beautifulsoup4